refactor: delete redundant types and files
