+++ src/extractors/youtube-new/parser/types/common.ts
@@ -59,3 +59,5 @@ export type ObjectPropertyRule  = ObjectRule & PropertyBase;
 type ArrayPropertyRule = ArrayRule & PropertyBase;
 
 export type PropertyRule = ArrayPropertyRule | ObjectPropertyRule | PrimitivePropertyRule;
+
+export type ObjectRuleProps<Rule extends ObjectRule> = Rule['properties'];
