+++ src/extractors/youtube-new/parser/types/remap.ts
@@ -1,5 +1,6 @@
-import type { ObjectRule } from "./common";
-import type { ObjectRuleProps } from "./props";
+import type { ObjectRule, ObjectRuleProps } from "./common";
+
+type RuleKeymap<Rule extends ObjectRule> = Rule['keymap']; 
 
 type RuleRemap<Rule extends ObjectRule> =
   Omit<Rule, 'properties'> & 
@@ -7,9 +8,9 @@ type RuleRemap<Rule extends ObjectRule> =
       properties: {
       [ 
         Key in keyof ObjectRuleProps<Rule> as 
-        Key extends keyof Rule['keymap'] ? 
-        Rule['keymap'][Key] extends string ? 
-        Rule['keymap'][Key] : 
+        Key extends keyof RuleKeymap<Rule> ? 
+        RuleKeymap<Rule>[Key] extends string ? 
+        RuleKeymap<Rule>[Key] : 
         Key : 
         Key
       ]: ObjectRuleProps<Rule>[Key] extends ObjectRule ? // Remap done recursively for flat to work correctly
